+++ nextjs/src/components/nav-bar/nav.ts
@@ -1,0 +1,5 @@
+export type NavItem = {
+  title: string
+  url?: string
+  items?: NavItem[]
+}