+++ nextjs/src/components/sections/section-card-wide.tsx
@@ -7,20 +7,28 @@ import clsx from "clsx"
 import { colors } from "@/lib/colors"
 
 type SectionMediaWideProps = {
-  reverse: boolean
+  reverse?: boolean
   children: React.ReactNode
-  image: {
+  /**
+   * @info if an image is passed, the component will use the image instead of childrenWide
+   */
+  image?: {
     src: string
     alt: string
   }
   color?: keyof typeof colors
-  ctas: CTA[]
+  ctas?: CTA[]
+  /**
+   * @info childrenWide is an alternative to image. It can be any component
+   */
+  childrenWide?: React.ReactNode
 }
 const SectionCardWide: React.FC<SectionMediaWideProps> = ({
   reverse,
   image,
   color,
   children,
+  childrenWide,
   ctas,
 }) => {
   return (
@@ -31,17 +39,20 @@ const SectionCardWide: React.FC<SectionMediaWideProps> = ({
       ])}
     >
       <span className="w-full lg:w-7/12">
-        <CardImage
-          className="w-full lg:max-h-[25vw]"
-          src={image.src}
-          alt={image.alt}
-          color={color}
-        />
+        {(image && (
+          <CardImage
+            className="w-full lg:max-h-[25vw]"
+            src={image.src}
+            alt={image.alt || ""}
+            color={color}
+          />
+        )) ||
+          childrenWide}
       </span>
 
-      <div className="w-full lg:w-5/12 grid gap-6 pt-8 lg:p-8 xl:p-12">
+      <div className="w-full lg:w-5/12 grid gap-6 pt-8 lg:p-8 xl:p-12 justify-items-start">
         {children}
-        {ctas.length > 0 && (
+        {ctas && ctas.length > 0 && (
           <div>
             <ButtonGroup ctas={ctas} />
           </div>