+++ nextjs/src/components/sections/section-whitepaper.tsx
@@ -39,9 +39,11 @@ const SectionWhitepaper: React.FC<SectionWhitepaperProps> = ({
           {title}
         </Title>
         <Text markdown={text} />
-        <LinkButton href={cta.href} variant={cta.variant} size={cta.size}>
-          {cta.text}
-        </LinkButton>
+        {cta.href && (
+          <LinkButton href={cta.href} variant={cta.variant} size={cta.size}>
+            {cta.text}
+          </LinkButton>
+        )}
       </div>
     </div>
   )