+++ nextjs/src/components/topbar.tsx
@@ -1,9 +1,14 @@
 "use client"
 import React, { useEffect } from "react"
-import NavMobile from "./nav-mobile"
-import NavDesktop from "./nav-desktop"
+import NavMobile from "./nav-bar/nav-mobile"
+import NavDesktop from "./nav-bar/nav-desktop"
+import { NavItem } from "./nav-bar/nav"
 
-const Topbar: React.FC = () => {
+type TopbarProps = {
+  navItems: NavItem[]
+}
+
+const Topbar: React.FC<TopbarProps> = ({ navItems }) => {
   return (
     <div>
       <div
@@ -11,8 +16,8 @@ const Topbar: React.FC = () => {
         id="navbar"
       >
         <div className="container mr-0">
-          <NavMobile />
-          <NavDesktop />
+          <NavMobile navItems={navItems} />
+          <NavDesktop navItems={navItems} />
         </div>
       </div>
     </div>